+++ src/pull_parser/any.rs
@@ -1,14 +1,10 @@
 //! Types and functions for all supported versions.
 
-use std::io::{Read, Seek};
+use std::io::{self, Seek};
 
 use crate::{
     low::{FbxHeader, FbxVersion},
-    pull_parser::{
-        self,
-        reader::{PlainSource, SeekableSource},
-        ParserSource, ParserVersion,
-    },
+    pull_parser::{self, ParserVersion},
 };
 
 pub use self::error::{Error, Result};
@@ -22,7 +18,50 @@ pub enum AnyParser<R> {
     V7400(pull_parser::v7400::Parser<R>),
 }
 
-impl<R: ParserSource> AnyParser<R> {
+impl<R: io::Read> AnyParser<R> {
+    /// Loads a tree from the given reader.
+    ///
+    /// This works for seekable readers (which implement [`std::io::Seek`]).
+    /// However, [`from_seekable_reader`][`Self::from_seekable_reader`] method
+    /// should be used for them because it is more efficent.
+    pub fn from_reader(mut reader: R) -> Result<Self> {
+        let header = FbxHeader::load(&mut reader)?;
+        match parser_version(header)? {
+            ParserVersion::V7400 => {
+                let parser = pull_parser::v7400::Parser::from_reader(header, reader)
+                    .unwrap_or_else(|e| {
+                        panic!(
+                    "Should never fail: FBX version {:?} should be supported by v7400 parser: {}",
+                    header.version(),
+                    e
+                )
+                    });
+                Ok(AnyParser::V7400(parser))
+            }
+        }
+    }
+
+    /// Loads a tree from the given seekable reader.
+    pub fn from_seekable_reader(mut reader: R) -> Result<Self>
+    where
+        R: Seek,
+    {
+        let header = FbxHeader::load(&mut reader)?;
+        match parser_version(header)? {
+            ParserVersion::V7400 => {
+                let parser = pull_parser::v7400::Parser::from_seekable_reader(header, reader)
+                    .unwrap_or_else(|e| {
+                        panic!(
+                    "Should never fail: FBX version {:?} should be supported by v7400 parser: {}",
+                    header.version(),
+                    e
+                )
+                    });
+                Ok(AnyParser::V7400(parser))
+            }
+        }
+    }
+
     /// Returns the parser version.
     #[inline]
     #[must_use]
@@ -47,43 +86,4 @@ fn parser_version(header: FbxHeader) -> Result<ParserVersion> {
     header
         .parser_version()
         .ok_or_else(|| Error::UnsupportedVersion(header.version()))
-}
-
-/// Loads a tree from the given reader.
-///
-/// This works for seekable readers (which implement [`std::io::Seek`]), but
-/// [`from_seekable_reader`] should be used for them, because it is more
-/// efficent.
-pub fn from_reader<R: Read>(mut reader: R) -> Result<AnyParser<PlainSource<R>>> {
-    let header = FbxHeader::load(&mut reader)?;
-    match parser_version(header)? {
-        ParserVersion::V7400 => {
-            let parser = pull_parser::v7400::from_reader(header, reader).unwrap_or_else(|e| {
-                panic!(
-                    "Should never fail: FBX version {:?} should be supported by v7400 parser: {}",
-                    header.version(),
-                    e
-                )
-            });
-            Ok(AnyParser::V7400(parser))
-        }
-    }
-}
-
-/// Loads a tree from the given seekable reader.
-pub fn from_seekable_reader<R: Read + Seek>(mut reader: R) -> Result<AnyParser<SeekableSource<R>>> {
-    let header = FbxHeader::load(&mut reader)?;
-    match parser_version(header)? {
-        ParserVersion::V7400 => {
-            let parser =
-                pull_parser::v7400::from_seekable_reader(header, reader).unwrap_or_else(|e| {
-                    panic!(
-                    "Should never fail: FBX version {:?} should be supported by v7400 parser: {}",
-                    header.version(),
-                    e
-                )
-                });
-            Ok(AnyParser::V7400(parser))
-        }
-    }
 }