+++ src/pull_parser/v7400/read.rs
@@ -2,9 +2,7 @@
 
 use std::io;
 
-use byteorder::{LittleEndian, ReadBytesExt};
-
-use crate::pull_parser::{v7400::Parser, ParserSource, Result};
+use crate::pull_parser::{v7400::Parser, Result};
 
 /// A trait for types readable from a reader.
 pub(crate) trait FromReader: Sized {
@@ -15,40 +13,44 @@ pub(crate) trait FromReader: Sized {
 impl FromReader for u8 {
     #[inline]
     fn from_reader(reader: &mut impl io::Read) -> Result<Self> {
-        Ok(reader.read_u8()?)
+        let mut buf = 0_u8;
+        reader.read_exact(std::slice::from_mut(&mut buf))?;
+        Ok(buf)
     }
 }
 
 /// Implement `FromReader` trait for primitive types.
 macro_rules! impl_from_reader_for_primitives {
-    ($ty:ty, $read_method:ident) => {
+    ($ty:ty) => {
         impl FromReader for $ty {
             #[inline]
             fn from_reader(reader: &mut impl io::Read) -> Result<Self> {
-                Ok(reader.$read_method::<LittleEndian>()?)
+                let mut buf = [0_u8; std::mem::size_of::<$ty>()];
+                reader.read_exact(&mut buf)?;
+                Ok(<$ty>::from_le_bytes(buf))
             }
         }
     };
 }
 
-impl_from_reader_for_primitives! { u16, read_u16 }
-impl_from_reader_for_primitives! { u32, read_u32 }
-impl_from_reader_for_primitives! { u64, read_u64 }
-impl_from_reader_for_primitives! { i16, read_i16 }
-impl_from_reader_for_primitives! { i32, read_i32 }
-impl_from_reader_for_primitives! { i64, read_i64 }
-impl_from_reader_for_primitives! { f32, read_f32 }
-impl_from_reader_for_primitives! { f64, read_f64 }
+impl_from_reader_for_primitives! { u16 }
+impl_from_reader_for_primitives! { u32 }
+impl_from_reader_for_primitives! { u64 }
+impl_from_reader_for_primitives! { i16 }
+impl_from_reader_for_primitives! { i32 }
+impl_from_reader_for_primitives! { i64 }
+impl_from_reader_for_primitives! { f32 }
+impl_from_reader_for_primitives! { f64 }
 
 /// A trait for types readable from a parser.
 pub(crate) trait FromParser: Sized {
     /// Reads the data from the given parser.
-    fn read_from_parser<R: ParserSource>(parser: &mut Parser<R>) -> Result<Self>;
+    fn read_from_parser<R: io::Read>(parser: &mut Parser<R>) -> Result<Self>;
 }
 
 impl<T: FromReader> FromParser for T {
     #[inline]
-    fn read_from_parser<R: ParserSource>(parser: &mut Parser<R>) -> Result<Self> {
+    fn read_from_parser<R: io::Read>(parser: &mut Parser<R>) -> Result<Self> {
         FromReader::from_reader(parser.reader())
     }
 }