+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 ## [Unreleased]
 
+## [0.9.0]
+
+* Remove `pull_parser::reader` module and items inside.
+    + Trait bounds `R: ParserSource` are now replaced with `R: std::io::Read`.
+* Change type parameters of `pull_parser::any::AnyParser` and `pull_parser::v7400::Parser`.
+* Remove free functions to create parsers, and add inherent methods instead.
+* Remove internal dependency to `byteorder` crate.
+
+### Changed (breaking)
+* Remove `pull_parser::reader` module and items inside.
+    + Trait bounds `R: ParserSource` are now replaced with `R: std::io::Read`.
+    + List of removed items:
+        - `pull_parser::reader` module
+        - `pull_parser::reader::ParserSource` trait
+        - `pull_parser::reader::PlainSource` type
+        - `pull_parser::reader::PositionCacheReader` type
+        - `pull_parser::reader::SeekableSource` type
+* Change type parameters of `pull_parser::any::AnyParser` and `pull_parser::v7400::Parser`.
+    + The parameter type of the parser is simplified. Previously parsers were
+      `{Any,}Parser<{Plain,Seekable}Source<R>>`, but now they are `{Any,}Parser<R>`.
+* Remove free functions to create parsers, and add inherent methods instead.
+    + List of removed functions and added alternatives:
+        - `pull_parser::any::from_reader()`: replaced with
+          `pull_parser::any::AnyParser::from_reader()`.
+        - `pull_parser::any::from_seekable_reader()`: replaced with
+          `pull_parser::any::AnyParser::from_seekable_reader()`.
+        - `pull_parser::v7400::from_reader()`: Replaced with
+        - `pull_parser::v7400::Parser::from_reader()`.
+        - `pull_parser::v7400::from_seekable_reader()`: Replaced with
+        - `pull_parser::v7400::Parser::from_seekable_reader()`.
+
+### Changed (non-breaking)
+* Remove internal dependency to `byteorder` crate.
+
 ## [0.8.2]
 
 * Add more tree manipulation methods.
@@ -339,7 +373,8 @@
 
 Totally rewritten.
 
-[Unreleased]: <https://github.com/lo48576/fbxcel/compare/v0.8.2...develop>
+[Unreleased]: <https://github.com/lo48576/fbxcel/compare/v0.9.0...develop>
+[0.9.0]: <https://github.com/lo48576/fbxcel/releases/tag/v0.9.0>
 [0.8.2]: <https://github.com/lo48576/fbxcel/releases/tag/v0.8.2>
 [0.8.1]: <https://github.com/lo48576/fbxcel/releases/tag/v0.8.1>
 [0.8.0]: <https://github.com/lo48576/fbxcel/releases/tag/v0.8.0>